fix icon anim playback, reset variations, improve automated sounds

+++ source/MugenHook/code/mugen/Sound.cpp
@@ -1,11 +1,12 @@
 #include "Sound.h"
 #include "..\mugenhook\base.h"
+#include "System.h"
 Sound * LoadSNDFile(const char * name)
 {
 	return CallAndReturn<Sound*, 0x46EF30, const char*>(name);
 }
 
-void PlaySound(Sound * snd, int group, int index, int unk, float vol)
+void PlaySound(Sound * snd, int group, int index, int unk, double vol)
 {
 	((void(__cdecl*)(Sound*, int, int, int, float))0x46ED30)(snd, group, index, unk, vol);
 }
